requirements.txt
matplotlib==3.4.3
numpy==1.21.1
opencv_python==4.5.4.58
pdf2image==1.16.0
Pillow==8.4.0
PyPDF2==1.26.0
